+++ frontend/src/app/components/find-customer/find-customer.component.ts
@@ -1,11 +1,60 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
+import { Router } from '@angular/router';
+import { FormsModule } from '@angular/forms';
+import { CustomerService } from '../../services/customer.service';
+import { CustomerDTO } from '../../../../models';
+import { JsonPipe, NgIf } from '@angular/common';
 
 @Component({
   selector: 'app-find-customer',
-  imports: [],
+  imports: [FormsModule, NgIf],
   templateUrl: './find-customer.component.html',
   styleUrl: './find-customer.component.css'
 })
 export class FindCustomerComponent {
 
+  router = inject(Router);
+  customerService = inject(CustomerService);
+
+  azonosito = 0;
+  szemelyiszam = "";
+  nev = "";
+
+  isTableShowing = false;
+
+  errorMessage: string = '';
+  customers: CustomerDTO[] = [];
+
+  openCustomerProfile(azonosito: number){
+    this.router.navigate(['customer-datasheet', azonosito]);
+  }
+
+  findCustomer() {
+    if (!this.azonosito && !this.szemelyiszam && !this.nev) {
+      this.errorMessage = 'Legalább egy keresési feltételt meg kell adni!';
+      return;
+    }
+
+    this.customerService.findCustomers(this.azonosito, this.szemelyiszam, this.nev).subscribe({
+      next: (customers) => {
+        if(Array.isArray(customers)){
+            this.isTableShowing = true;
+            this.customers = customers
+
+        } else {
+          this.customers = [customers];
+          this.isTableShowing = true;
+        }
+        this.errorMessage = '';
+      },
+      error: (err) => {
+        this.errorMessage = err.error.message;
+      }
+    });
+
+  }
+
+  openMainMenu(){
+    this.router.navigate(['main-menu']);
+  }
 }